requirement.txt
streamlit
pandas
feedparser
python-dateutil
openai